+++ CHANGES.md
@@ -93,7 +93,14 @@ next
 
 - Make `dev` the default build profile (#920, @diml)
 
-- Version `dune-workspace` and `~/.config/dune/config` files (#..., @diml)
+- Version `dune-workspace` and `~/.config/dune/config` files (#932, @diml)
+
+- Add the ability to build an alias non-recursively from the command
+  line by writing `@@alias` (#926, @diml)
+
+- Add a special `default` alias that defaults to `(alias_rec install)`
+  when not defined by the user and make `@@default` be the default
+  target (#926, @diml)
 
 - Add `%{profile}` variable. (#938, @rgrinberg)